Just some minor changes; committing now since I haven't done so in a while and want to have a recent backup.

+++ src/GameManager.h
@@ -60,7 +60,7 @@ class GameManager : public IEventReceiver {
 		bool OnEvent( const SEvent& );
 
 		void resetThings();
-		int run();
+		uint_fast8_t run();
 
 		void setControls();
